requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
fastapi>=0.110.0   # optional if you want HTTP endpoint
uvicorn>=0.30.0    # to run FastAPI
openai-agents>=0.2.8
google-generativeai>=0.7.2